scDotPlot

Plot DotPlot from Seurat object

Usage

scDotPlot(
  x,
  group.by,
  split.by = NULL,
  features,
  assay = "RNA",
  slot = "data",
  scale = T,
  palette = "Reds",
  direction = 1,
  diffexp = NULL,
  coord_flip = T,
  unique = F
)

Arguments

x

Seurat object

group.by

metadata column to group by

split.by

metadata column to group by

features

feature names stored in a vector or in a named list

assay

assay name, defaults to "RNA"

slot

slot name, defaults to "data"

scale

if TRUE, scale gene expression across cells. Defaults to TRUE

palette

a vector of colors, defaults to "Reds" from RColorBrewer

direction

direction of palette, default to 1

diffexp

whether to plot significance, input can be either a dataframe or a csv file path of the Seurat::FindAllMarkers output. Defaults to NULL.

coord_flip

If TRUE, plot features on y-axis. Defaults to TRUE

unique

If TRUE, do not repeat feature names. Defaults to FALSE